+++ src/invoices/InvoicesList/InvoicesListFilters/InvoicesListFilters.js
@@ -27,6 +27,7 @@ const applyFiltersAdapter = (applyFilters) => ({ name, values }) => applyFilters
 const InvoicesListFilters = ({
   activeFilters,
   applyFilters,
+  disabled,
 }) => {
   const adaptedApplyFilters = useCallback(
     applyFiltersAdapter(applyFilters),
@@ -42,11 +43,13 @@ const InvoicesListFilters = ({
         name={FILTERS.STATUS}
         onChange={adaptedApplyFilters}
         options={INVOICE_STATUSES_OPTIONS}
+        disabled={disabled}
       />
 
       <PluggableOrganizationFilter
         id={FILTERS.VENDOR}
         activeFilters={activeFilters[FILTERS.VENDOR]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.vendorName"
         name={FILTERS.VENDOR}
         onChange={adaptedApplyFilters}
@@ -55,6 +58,7 @@ const InvoicesListFilters = ({
       <AcqDateRangeFilter
         id={FILTERS.DATE_CREATED}
         activeFilters={activeFilters[FILTERS.DATE_CREATED]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.dateCreated"
         name={FILTERS.DATE_CREATED}
         onChange={adaptedApplyFilters}
@@ -62,6 +66,7 @@ const InvoicesListFilters = ({
       <AcqDateRangeFilter
         id={FILTERS.INVOICE_DATE}
         activeFilters={activeFilters[FILTERS.INVOICE_DATE]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.details.information.invoiceDate"
         name={FILTERS.INVOICE_DATE}
         onChange={adaptedApplyFilters}
@@ -69,12 +74,14 @@ const InvoicesListFilters = ({
       <AcqUnitFilter
         id={FILTERS.ACQUISITIONS_UNIT}
         activeFilters={activeFilters[FILTERS.ACQUISITIONS_UNIT]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.acquisitionsUnit"
         name={FILTERS.ACQUISITIONS_UNIT}
         onChange={adaptedApplyFilters}
       />
       <AcqTagsFilter
         activeFilters={activeFilters[FILTERS.TAGS]}
+        disabled={disabled}
         id={FILTERS.TAGS}
         name={FILTERS.TAGS}
         onChange={adaptedApplyFilters}
@@ -82,6 +89,7 @@ const InvoicesListFilters = ({
       <AcqDateRangeFilter
         id={FILTERS.PAYMENT_DUE}
         activeFilters={activeFilters[FILTERS.PAYMENT_DUE]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.details.information.paymentDue"
         name={FILTERS.PAYMENT_DUE}
         onChange={adaptedApplyFilters}
@@ -89,6 +97,7 @@ const InvoicesListFilters = ({
       <AcqCheckboxFilter
         id={FILTERS.PAYMENT_METHOD}
         activeFilters={activeFilters[FILTERS.PAYMENT_METHOD]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.paymentMethod"
         name={FILTERS.PAYMENT_METHOD}
         onChange={adaptedApplyFilters}
@@ -97,18 +106,21 @@ const InvoicesListFilters = ({
       <AcqDateRangeFilter
         id={FILTERS.APPROVAL_DATE}
         activeFilters={activeFilters[FILTERS.APPROVAL_DATE]}
+        disabled={disabled}
         labelId="ui-invoice.invoice.approvalDate"
         name={FILTERS.APPROVAL_DATE}
         onChange={adaptedApplyFilters}
       />
       <SourceFilter
         activeFilters={activeFilters[FILTERS.SOURCE]}
+        disabled={disabled}
         id={FILTERS.SOURCE}
         name={FILTERS.SOURCE}
         onChange={adaptedApplyFilters}
       />
       <BooleanFilter
         activeFilters={activeFilters[FILTERS.EXPORT_TO_ACCOUNTING]}
+        disabled={disabled}
         id={FILTERS.EXPORT_TO_ACCOUNTING}
         labelId="ui-invoice.invoice.exportToAccounting"
         name={FILTERS.EXPORT_TO_ACCOUNTING}
@@ -121,6 +133,7 @@ const InvoicesListFilters = ({
 InvoicesListFilters.propTypes = {
   activeFilters: PropTypes.object.isRequired,
   applyFilters: PropTypes.func.isRequired,
+  disabled: PropTypes.bool.isRequired,
 };
 
 export default InvoicesListFilters;